services/ssg/requirements.txt
psycopg2-binary==2.9.10
requests==2.32.3
typing-inspection==0.4.0
typing_extensions==4.13.0
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Flask==3.1.0
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Werkzeug==3.1.3
Markdown==3.7
google-crc32c==1.7.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-resumable-media==2.7.2
SQLAlchemy==2.0.40
greenlet==3.1.1
pg8000==1.31.2
aiofiles==24.1.0
aiohttp==3.11.17
cloud-sql-python-connector==1.18.1
cryptography==44.0.2
dnspython==2.7.0